backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.4
anthropic==0.28.0
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
supabase==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sse-starlette==2.1.2
